% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_label.R
\name{generate_label}
\alias{generate_label}
\title{Function generate_label()}
\usage{
generate_label(summary_entity, model = "gpt-3.5-turbo", get_cluster = TRUE)
}
\arguments{
\item{model}{variable containing the version of ChatGPT to be used}

\item{dataframe}{dataframe containing information about tiktok accounts}

\item{get_cluste}{boolean variable that specifies whether to also analyze clusters}
}
\description{
This function allows you to make a request to ChatGPT which, processing the data taken as input,
returns a summary label of the descriptions of the videos that are part of each cluster or component.
}
\examples{
generate_label(dataframe = dataframe, model = "gpt-3.5-turbo", get_cluster = TRUE)
}
