% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiktok_account_info.R
\name{tiktok_account_info}
\alias{tiktok_account_info}
\title{tiktok_account_info function}
\arguments{
\item{dataframe}{the result of account_stats function from CoorTweet library}

\item{summary_entity}{the result of create_entity function}
}
\description{
This function add information about the accounts that are detected for their coordinated behaviours
}
\examples{
tiktok_account_info(summary_accounts, summary_entity)
}
